refactor(core): split vm lifecycle types
