+++ app/arcbox-core/src/vm_lifecycle/mod.rs
@@ -31,11 +31,14 @@ mod health;
 mod recovery;
 #[cfg(target_os = "macos")]
 mod serial;
+#[cfg(test)]
+mod tests;
+mod types;
 
 use crate::boot_assets::BootAssetProvider;
 use crate::error::{CoreError, Result};
 use crate::event::{Event, EventBus};
-use crate::machine::{MachineConfig, MachineInfo, MachineManager, MachineState};
+use crate::machine::{MachineConfig, MachineInfo, MachineManager};
 use arcbox_constants::cmdline::{GUEST_DOCKER_VSOCK_PORT_KEY, HV_EARLYCON_DIRECTIVE};
 use arcbox_error::CommonError;
 use std::fs::OpenOptions;
@@ -80,223 +83,12 @@ const DOCKER_DATA_IMAGE_NAME: &str = "docker.img";
 /// and prevents users from hitting artificial limits.
 const DOCKER_DATA_IMAGE_SIZE_BYTES: u64 = 8 * 1024 * 1024 * 1024 * 1024;
 
-/// Extended VM lifecycle state.
-///
-/// This extends the basic `MachineState` with additional states
-/// for lifecycle management.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum VmLifecycleState {
-    /// VM does not exist yet.
-    NotExist,
-    /// VM is being created.
-    Creating,
-    /// VM created but not started.
-    Created,
-    /// VM is starting up.
-    Starting,
-    /// VM is running and agent is ready.
-    Running,
-    /// VM is idle (no recent activity).
-    Idle,
-    /// VM is stopping.
-    Stopping,
-    /// VM has stopped.
-    Stopped,
-    /// VM failed to start or crashed.
-    Failed,
-}
-
-impl VmLifecycleState {
-    /// Returns true if VM is in a state where it can accept commands.
-    #[must_use]
-    pub const fn is_ready(&self) -> bool {
-        matches!(self, Self::Running | Self::Idle)
-    }
-
-    /// Returns true if VM needs to be started.
-    #[must_use]
-    pub const fn needs_start(&self) -> bool {
-        matches!(
-            self,
-            Self::NotExist | Self::Created | Self::Stopped | Self::Failed
-        )
-    }
-
-    /// Returns the state name for logging.
-    #[must_use]
-    pub const fn as_str(&self) -> &'static str {
-        match self {
-            Self::NotExist => "not_exist",
-            Self::Creating => "creating",
-            Self::Created => "created",
-            Self::Starting => "starting",
-            Self::Running => "running",
-            Self::Idle => "idle",
-            Self::Stopping => "stopping",
-            Self::Stopped => "stopped",
-            Self::Failed => "failed",
-        }
-    }
-}
-
-impl From<MachineState> for VmLifecycleState {
-    fn from(state: MachineState) -> Self {
-        match state {
-            MachineState::Created => Self::Created,
-            MachineState::Starting => Self::Starting,
-            MachineState::Running => Self::Running,
-            MachineState::Stopping => Self::Stopping,
-            MachineState::Stopped => Self::Stopped,
-        }
-    }
-}
-
-/// State transition events.
-#[derive(Debug, Clone)]
-pub enum VmEvent {
-    /// Request to create VM.
-    Create,
-    /// VM creation completed.
-    Created,
-    /// Request to start VM.
-    Start,
-    /// Agent became ready.
-    AgentReady,
-    /// VM became idle (no activity for `idle_timeout`).
-    IdleTimeout,
-    /// Activity detected, exit idle state.
-    Activity,
-    /// Request to stop VM.
-    Stop,
-    /// VM stopped successfully.
-    Stopped,
-    /// Force stop VM.
-    ForceStop,
-    /// VM crashed or failed.
-    Failure(String),
-    /// Retry after failure.
-    Retry,
-}
-
-/// VM lifecycle configuration.
-#[derive(Debug, Clone)]
-pub struct VmLifecycleConfig {
-    /// Enable auto-stop after idle timeout.
-    pub auto_stop: bool,
-    /// Idle timeout before entering idle state.
-    pub idle_timeout: Duration,
-    /// Startup timeout for VM boot.
-    pub startup_timeout: Duration,
-    /// Health check interval.
-    pub health_check_interval: Duration,
-    /// Maximum retry attempts for recovery.
-    pub max_retries: u32,
-    /// Default VM configuration.
-    pub default_vm: DefaultVmConfig,
-    /// Skip VM check (for testing only).
-    /// When true, `ensure_ready()` returns immediately with a mock CID.
-    pub skip_vm_check: bool,
-    /// Guest docker API vsock port propagated via kernel cmdline.
-    pub guest_docker_vsock_port: Option<u32>,
-    /// macOS hypervisor backend this lifecycle's machine will use.
-    ///
-    /// Defaults to `Hv` so the native utility VM keeps running on the
-    /// custom HV-framework path. The rosetta utility VM overrides this
-    /// to `Vz` so it can host the Rosetta share.
-    pub backend: arcbox_vmm::VmBackend,
-}
-
-impl Default for VmLifecycleConfig {
-    fn default() -> Self {
-        Self {
-            auto_stop: true,
-            idle_timeout: Duration::from_secs(DEFAULT_IDLE_TIMEOUT_SECS),
-            startup_timeout: Duration::from_secs(DEFAULT_STARTUP_TIMEOUT_SECS),
-            health_check_interval: Duration::from_secs(DEFAULT_HEALTH_CHECK_INTERVAL_SECS),
-            max_retries: DEFAULT_MAX_RETRIES,
-            default_vm: DefaultVmConfig::default(),
-            skip_vm_check: false,
-            guest_docker_vsock_port: None,
-            backend: arcbox_vmm::VmBackend::Hv,
-        }
-    }
-}
-
-/// Default VM configuration.
-#[derive(Debug, Clone)]
-pub struct DefaultVmConfig {
-    /// Number of vCPUs (default: host core count).
-    pub cpus: u32,
-    /// Memory in MB (default: half of host RAM, clamped to 512–16384).
-    pub memory_mb: u64,
-    /// Disk size in GB (default: 50).
-    pub disk_gb: u64,
-    /// Path to kernel image (if None, use `BootAssetProvider`).
-    pub kernel: Option<PathBuf>,
-    /// Kernel command line.
-    pub cmdline: Option<String>,
-    /// Enable Rosetta for x86 emulation (Apple Silicon only).
-    pub rosetta: bool,
-}
-
-impl Default for DefaultVmConfig {
-    fn default() -> Self {
-        Self {
-            cpus: arcbox_hypervisor::default_vm_cpu_count(),
-            memory_mb: arcbox_hypervisor::default_vm_memory_size() / (1024 * 1024),
-            disk_gb: 50,
-            kernel: None,
-            cmdline: None,
-            rosetta: cfg!(target_arch = "aarch64"),
-        }
-    }
-}
-
-/// Boot parameters for the default VM, resolved from config + boot assets.
-///
-/// Produced by [`VmLifecycleManager::resolve_desired_boot`] and consumed both
-/// when creating the machine and when checking an existing machine for drift,
-/// so the two paths can never disagree about the desired kernel/cmdline.
-struct DesiredBoot {
-    /// Resolved kernel image path.
-    kernel: String,
-    /// Final kernel command line (after quiet-strip, earlycon, and vsock port).
-    cmdline: String,
-    /// EROFS rootfs image path (read-only vda).
-    rootfs_image: PathBuf,
-}
-
-/// Returns the first daemon-overridable field that differs between a persisted
-/// machine and the desired default-VM config, or `None` if it is up to date.
-///
-/// Centralizing drift detection keeps every overridable field checked in one
-/// place and in sync with what [`VmLifecycleManager::create_default_machine`]
-/// produces: kernel and cmdline are compared against the same resolved
-/// [`DesiredBoot`] the machine would be created with, so a change to `--kernel`
-/// or any cmdline-injected override (e.g. the guest docker vsock port) forces a
-/// recreate instead of silently reusing a stale VM.
-fn machine_drift_reason(
-    persisted: &MachineInfo,
-    want: &DefaultVmConfig,
-    boot: &DesiredBoot,
-) -> Option<&'static str> {
-    if persisted.cpus != want.cpus {
-        Some("cpus")
-    } else if persisted.memory_mb != want.memory_mb {
-        Some("memory_mb")
-    } else if persisted.kernel.as_deref() != Some(boot.kernel.as_str()) {
-        Some("kernel")
-    } else if persisted.cmdline.as_deref() != Some(boot.cmdline.as_str()) {
-        Some("cmdline")
-    } else {
-        None
-    }
-}
-
 // Note: BootAssetProvider and BootAssets are now in crate::boot_assets module.
 
 pub use health::HealthMonitor;
 pub use recovery::{BackoffStrategy, RecoveryAction, RecoveryPolicy};
+pub use types::{DefaultVmConfig, VmEvent, VmLifecycleConfig, VmLifecycleState};
+use types::{DesiredBoot, machine_drift_reason};
 
 /// VM lifecycle manager.
 ///
@@ -1279,204 +1071,4 @@ impl VmLifecycleManager {
 
 const fn is_not_found_error(err: &CoreError) -> bool {
     matches!(err, CoreError::Common(CommonError::NotFound(_)))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_lifecycle_state_is_ready() {
-        assert!(!VmLifecycleState::NotExist.is_ready());
-        assert!(!VmLifecycleState::Creating.is_ready());
-        assert!(!VmLifecycleState::Created.is_ready());
-        assert!(!VmLifecycleState::Starting.is_ready());
-        assert!(VmLifecycleState::Running.is_ready());
-        assert!(VmLifecycleState::Idle.is_ready());
-        assert!(!VmLifecycleState::Stopping.is_ready());
-        assert!(!VmLifecycleState::Stopped.is_ready());
-        assert!(!VmLifecycleState::Failed.is_ready());
-    }
-
-    #[test]
-    fn test_lifecycle_state_needs_start() {
-        assert!(VmLifecycleState::NotExist.needs_start());
-        assert!(!VmLifecycleState::Creating.needs_start());
-        assert!(VmLifecycleState::Created.needs_start());
-        assert!(!VmLifecycleState::Starting.needs_start());
-        assert!(!VmLifecycleState::Running.needs_start());
-        assert!(!VmLifecycleState::Idle.needs_start());
-        assert!(!VmLifecycleState::Stopping.needs_start());
-        assert!(VmLifecycleState::Stopped.needs_start());
-        assert!(VmLifecycleState::Failed.needs_start());
-    }
-
-    #[test]
-    fn ensure_earlycon_upgrades_bare_on_hv() {
-        let out = ensure_earlycon(
-            "console=hvc0 earlycon root=/dev/vda".to_string(),
-            arcbox_vmm::VmBackend::Hv,
-        );
-        assert!(out.contains(HV_EARLYCON_DIRECTIVE));
-        // The bare token is replaced, not left dangling alongside the directive.
-        assert!(!out.split_whitespace().any(|t| t == "earlycon"));
-    }
-
-    #[test]
-    fn ensure_earlycon_leaves_vz_untouched() {
-        // VZ has no PL011 device; the cmdline must pass through verbatim.
-        let cmdline = "console=hvc0 earlycon root=/dev/vda".to_string();
-        let out = ensure_earlycon(cmdline.clone(), arcbox_vmm::VmBackend::Vz);
-        assert_eq!(out, cmdline);
-        assert!(!out.contains("pl011"));
-    }
-
-    #[test]
-    fn ensure_earlycon_respects_explicit_directive() {
-        let cmdline = "console=hvc0 earlycon=pl011,0x9000000 root=/dev/vda".to_string();
-        let out = ensure_earlycon(cmdline.clone(), arcbox_vmm::VmBackend::Hv);
-        assert_eq!(out, cmdline);
-    }
-
-    #[test]
-    fn agent_timeout_error_folds_last_error() {
-        let bare = agent_timeout_error(None).to_string();
-        assert!(bare.contains("timeout waiting for agent"));
-        assert!(!bare.contains("last error"));
-
-        let folded = agent_timeout_error(Some("guest reported: disk full")).to_string();
-        assert!(folded.contains("timeout waiting for agent"));
-        assert!(folded.contains("guest reported: disk full"));
-    }
-
-    #[test]
-    fn test_default_config() {
-        let config = VmLifecycleConfig::default();
-        assert!(config.auto_stop);
-        assert_eq!(config.max_retries, DEFAULT_MAX_RETRIES);
-    }
-
-    #[test]
-    fn ensure_sparse_block_image_is_thin_provisioned() {
-        // Regression guard (issue #244 / ABXD-95): a freshly created data image
-        // must be thin — report its full virtual size yet consume virtually no
-        // physical disk. A reintroduced upfront preallocation (e.g. macOS
-        // `F_PREALLOCATE`) would balloon the allocated block count and fail here.
-        let dir = tempfile::tempdir().unwrap();
-        // Nested path also exercises parent-directory creation.
-        let path = dir.path().join("data").join("docker.img");
-        let size_bytes = DOCKER_DATA_IMAGE_SIZE_BYTES; // 8 TiB virtual size
-
-        VmLifecycleManager::ensure_sparse_block_image(&path, size_bytes).unwrap();
-
-        let meta = std::fs::metadata(&path).unwrap();
-        assert_eq!(
-            meta.len(),
-            size_bytes,
-            "logical size must match the requested virtual size"
-        );
-
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::MetadataExt;
-            // `blocks()` counts 512-byte blocks actually allocated on disk. A
-            // sparse file reserves none up front; 1 MiB of slack covers any
-            // filesystem metadata overhead while still catching a multi-GiB
-            // preallocation regression.
-            let physical_bytes = meta.blocks() * 512;
-            assert!(
-                physical_bytes < 1024 * 1024,
-                "image must be sparse: {physical_bytes} physical bytes allocated \
-                 for an empty {size_bytes}-byte image"
-            );
-        }
-    }
-
-    #[test]
-    fn ensure_sparse_block_image_never_shrinks() {
-        let dir = tempfile::tempdir().unwrap();
-        let path = dir.path().join("docker.img");
-
-        VmLifecycleManager::ensure_sparse_block_image(&path, 8192).unwrap();
-        assert_eq!(std::fs::metadata(&path).unwrap().len(), 8192);
-
-        // A call with a smaller virtual size must not truncate the existing
-        // image — that would discard guest data.
-        VmLifecycleManager::ensure_sparse_block_image(&path, 4096).unwrap();
-        assert_eq!(std::fs::metadata(&path).unwrap().len(), 8192);
-
-        // A larger size grows it.
-        VmLifecycleManager::ensure_sparse_block_image(&path, 16384).unwrap();
-        assert_eq!(std::fs::metadata(&path).unwrap().len(), 16384);
-    }
-
-    #[test]
-    fn test_default_vm_config() {
-        let config = DefaultVmConfig::default();
-        assert_eq!(config.cpus, arcbox_hypervisor::default_vm_cpu_count());
-        // Default memory is half of host RAM, clamped to [512, 16384] MB.
-        let expected_mb = arcbox_hypervisor::default_vm_memory_size() / (1024 * 1024);
-        assert_eq!(config.memory_mb, expected_mb);
-        assert!(config.memory_mb >= 512);
-        assert!(config.memory_mb <= 16384);
-        assert_eq!(config.disk_gb, 50);
-    }
-
-    fn sample_machine(cpus: u32, memory_mb: u64, kernel: &str, cmdline: &str) -> MachineInfo {
-        MachineInfo {
-            name: "default".to_string(),
-            state: MachineState::Created,
-            vm_id: crate::vm::VmId::new(),
-            cid: None,
-            cpus,
-            memory_mb,
-            disk_gb: 50,
-            kernel: Some(kernel.to_string()),
-            cmdline: Some(cmdline.to_string()),
-            block_devices: Vec::new(),
-            distro: None,
-            distro_version: None,
-            disk_path: None,
-            ssh_key_path: None,
-            ip_address: None,
-            created_at: chrono::Utc::now(),
-        }
-    }
-
-    #[test]
-    fn machine_drift_detects_each_overridable_field() {
-        let want = DefaultVmConfig {
-            cpus: 4,
-            memory_mb: 4096,
-            ..DefaultVmConfig::default()
-        };
-        let boot = DesiredBoot {
-            kernel: "/k".to_string(),
-            cmdline: "console=hvc0 earlycon".to_string(),
-            rootfs_image: std::path::PathBuf::from("/rootfs.erofs"),
-        };
-        let current = sample_machine(want.cpus, want.memory_mb, &boot.kernel, &boot.cmdline);
-
-        // Matching machine: no drift.
-        assert_eq!(machine_drift_reason(&current, &want, &boot), None);
-
-        // Each overridable field, changed independently, is detected.
-        let mut m = current.clone();
-        m.cpus = want.cpus + 1;
-        assert_eq!(machine_drift_reason(&m, &want, &boot), Some("cpus"));
-
-        let mut m = current.clone();
-        m.memory_mb = want.memory_mb + 1;
-        assert_eq!(machine_drift_reason(&m, &want, &boot), Some("memory_mb"));
-
-        let mut m = current.clone();
-        m.kernel = Some("/other-kernel".to_string());
-        assert_eq!(machine_drift_reason(&m, &want, &boot), Some("kernel"));
-
-        // The cmdline gap that previously slipped through (e.g. arm64.nosve
-        // added/removed without bumping the boot-asset version).
-        let mut m = current;
-        m.cmdline = Some("console=hvc0 earlycon arm64.nosve".to_string());
-        assert_eq!(machine_drift_reason(&m, &want, &boot), Some("cmdline"));
-    }
 }
